Fix ad policy property being in the base data trait
